refactor(server): establish chat execution coordinator
