+++ server/__tests__/queue-transcript-stability.test.js
@@ -3,7 +3,7 @@ import fs from 'fs/promises';
 import os from 'os';
 import path from 'path';
 import { UserMessage } from '../../common/chat-types.js';
-import { QueueManager } from '../queue.js';
+import { ChatExecutionCoordinator } from '../chat-execution/chat-execution-coordinator.js';
 import { ChatExecutionActivity } from '../chats/chat-execution-activity.js';
 import { ChatNativeReloader } from '../chats/chat-native-reload.js';
 import { ChatRunningError } from '../chats/errors.js';
@@ -68,7 +68,7 @@ describe('queue and transcript stability', () => {
     const runtime = new HeldDirectRuntime();
     let agentSessionId = null;
     try {
-      const queue = new QueueManager(
+      const queue = new ChatExecutionCoordinator(
         workspaceDir,
         {
           runAgentTurn: mock(async () => undefined),
@@ -167,7 +167,7 @@ describe('queue and transcript stability', () => {
       const coordinator = new UserAbortLifecycleCoordinator(pendingInputs, {
         terminalTimeoutMs: 60_000,
       });
-      const queue = new QueueManager(
+      const queue = new ChatExecutionCoordinator(
         workspaceDir,
         {
           runAgentTurn: mock(async (_chatId, content, options) => {
@@ -273,7 +273,7 @@ describe('queue and transcript stability', () => {
         isChatRunning: mock(() => false),
         waitUntilTurnAbortable: mock(() => runtimeAbortable.promise),
       };
-      const queue = new QueueManager(
+      const queue = new ChatExecutionCoordinator(
         workspaceDir,
         turnRunner,
         pendingPort,
@@ -341,7 +341,7 @@ describe('queue and transcript stability', () => {
         ),
       });
       let turnCount = 0;
-      const queue = new QueueManager(
+      const queue = new ChatExecutionCoordinator(
         workspaceDir,
         {
           runAgentTurn: mock(async (_chatId, content, options) => {
@@ -444,7 +444,7 @@ describe('queue and transcript stability', () => {
         () => ({}),
         () => true,
       ];
-      const queue = new QueueManager(workspaceDir, ...queueDeps);
+      const queue = new ChatExecutionCoordinator(workspaceDir, ...queueDeps);
       const created = await queue.createChatQueueEntry(chatId, 'survive restart');
       await queue.popNextChat(chatId);
 
@@ -459,7 +459,7 @@ describe('queue and transcript stability', () => {
       expect(accepted.kind).toBe('accepted');
       await ledger.update(accepted.record.key, { status: 'scheduled' });
 
-      const restartedQueue = new QueueManager(workspaceDir, ...queueDeps);
+      const restartedQueue = new ChatExecutionCoordinator(workspaceDir, ...queueDeps);
       await restartedQueue.recoverChatExecutionControls();
       const recoveredQueue = await restartedQueue.readChatExecutionControl(chatId);
       expect(recoveredQueue.entries).toMatchObject([{
