+++ server/chat-execution/types.ts
@@ -1,0 +1,309 @@
+import crypto from 'crypto';
+import type { AutomaticQueuePauseKind, QueueEntry } from '../../common/queue-state.ts';
+import type {
+  ChatImage,
+  ChatMessage,
+  ChatStopIntent,
+  UserMessageDeliveryStatus,
+} from '../../common/chat-types.ts';
+import type { ChatViewMessage } from '../../common/chat-view.ts';
+import type { AgentExecutionAdmission, RunAgentTurnOptions } from '../agents/session-types.ts';
+import {
+  cloneStoredChatExecutionControl,
+  type StoredChatExecutionControlState,
+} from '../chat-execution-control-state.ts';
+import { DomainError } from '../lib/domain-error.ts';
+import type { TurnIdentity } from '../lib/turn-identity.ts';
+import type { QueuedTurnFinalizationOutcome } from './turn-finalization-tracker.ts';
+import type {
+  QueueCommandIdentity,
+  ReceiptRetention,
+  TransitionContext,
+  TransitionRejection,
+} from './chat-execution-control-transitions.ts';
+
+export type PendingUserInputRegistrationOptions = Pick<
+  RunAgentTurnOptions,
+  'clientRequestId' | 'clientMessageId' | 'turnId' | 'images'
+> & {
+  deliveryStatus?: UserMessageDeliveryStatus;
+};
+
+export class QueueEntryMutationError extends DomainError {
+  readonly control: StoredChatExecutionControlState;
+
+  constructor(
+    code: 'QUEUE_ENTRY_NOT_FOUND' | 'QUEUE_ENTRY_ALREADY_SENT' | 'QUEUE_ENTRY_REVISION_CONFLICT',
+    message: string,
+    control: StoredChatExecutionControlState,
+  ) {
+    super(code, message, code === 'QUEUE_ENTRY_NOT_FOUND' ? 404 : 409);
+    this.name = 'QueueEntryMutationError';
+    this.control = cloneStoredChatExecutionControl(control);
+  }
+}
+
+export class QueuePauseChangedError extends DomainError {
+  readonly control: StoredChatExecutionControlState;
+
+  constructor(control: StoredChatExecutionControlState) {
+    super('QUEUE_PAUSE_CHANGED', 'The queue pause changed before it could be resumed', 409);
+    this.name = 'QueuePauseChangedError';
+    this.control = cloneStoredChatExecutionControl(control);
+  }
+}
+
+export class RecoveredInputContinuationChangedError extends DomainError {
+  readonly control: StoredChatExecutionControlState;
+
+  constructor(control: StoredChatExecutionControlState) {
+    super(
+      'RECOVERED_INPUT_CONTINUATION_CHANGED',
+      'The recovered-input continuation changed before it could be applied',
+      409,
+      true,
+    );
+    this.name = 'RecoveredInputContinuationChangedError';
+    this.control = cloneStoredChatExecutionControl(control);
+  }
+}
+
+export class RecoveredInputContinuationRequiresQueueError extends DomainError {
+  readonly control: StoredChatExecutionControlState;
+
+  constructor(control: StoredChatExecutionControlState) {
+    super(
+      'RECOVERED_INPUT_CONTINUATION_REQUIRES_QUEUE',
+      'Continue queue requires at least one queued message',
+      409,
+      true,
+    );
+    this.name = 'RecoveredInputContinuationRequiresQueueError';
+    this.control = cloneStoredChatExecutionControl(control);
+  }
+}
+
+export interface QueueCommandMutationResult {
+  entryId: string;
+  control: StoredChatExecutionControlState;
+  duplicate: boolean;
+}
+
+export interface StopActiveTurnResult {
+  stopped: boolean;
+  control: StoredChatExecutionControlState;
+}
+
+export interface DirectTurnReservation {
+  readonly chatId: string;
+  readonly reservationId: string;
+  readonly executionAdmission: AgentExecutionAdmission;
+}
+
+export interface AgentTurnRunnerPort {
+  runAgentTurn(chatId: string, command: string, options: RunAgentTurnOptions): Promise<void>;
+  submitActiveInput?(
+    chatId: string,
+    command: string,
+    options: RunAgentTurnOptions,
+    beforeDelivery: () => Promise<void>,
+  ): Promise<boolean>;
+  abortSession(chatId: string): Promise<boolean>;
+  isChatRunning(chatId: string): boolean;
+  waitUntilTurnAbortable(
+    chatId: string,
+    turn: TurnIdentity,
+    signal?: AbortSignal,
+  ): Promise<boolean>;
+}
+
+export interface PendingInputsPort {
+  register(
+    chatId: string,
+    content: string,
+    options?: {
+      clientRequestId?: string;
+      clientMessageId?: string;
+      turnId?: string;
+      images?: ChatImage[];
+      deliveryStatus?: UserMessageDeliveryStatus;
+    },
+  ): Promise<unknown>;
+  discard(chatId: string, clientRequestId: string): boolean;
+  markFailed(chatId: string, clientRequestId: string): boolean;
+  markUnconfirmed(chatId: string, clientRequestId: string): boolean;
+}
+
+export interface ChatMessagesPort {
+  appendMessages(
+    chatId: string,
+    messages: ChatMessage[],
+  ): Promise<{ generationId: string; messages: ChatViewMessage[] }>;
+}
+
+export type ExecutionControlUpdatedCallback = (
+  chatId: string,
+  control: StoredChatExecutionControlState,
+) => void;
+export type DispatchingCallback = (chatId: string, entryId: string, content: string) => void;
+export type SessionStopRequestedCallback = (
+  chatId: string,
+  stopId: string,
+  turn: TurnIdentity | undefined,
+) => void;
+export type SessionStoppedCallback = (
+  chatId: string,
+  success: boolean,
+  intent: ChatStopIntent,
+  stopId: string,
+) => void;
+export type ChatIdleCallback = (chatId: string) => void;
+export type TurnFailedCallback = (
+  chatId: string,
+  errorMessage: string,
+  options: RunAgentTurnOptions,
+) => void;
+export type TurnSettledCallback = (chatId: string, turn: TurnIdentity | undefined) => void;
+export type ChatMessagesCallback = (
+  chatId: string,
+  generationId: string,
+  messages: ChatViewMessage[],
+  metadata?: { clientRequestId?: string; turnId?: string },
+) => void;
+export type QueueDrainOptionsResolver = (chatId: string) => RunAgentTurnOptions;
+export type ChatExistsResolver = (chatId: string) => boolean;
+
+export interface SessionStopInFlight {
+  intent: ChatStopIntent;
+  stopId: string;
+  promise: Promise<boolean>;
+  resolve(success: boolean): void;
+  reject(error: unknown): void;
+  started: boolean;
+}
+
+export type DrainSuppressionReason = 'abort' | 'manual-stop' | 'deletion';
+
+export interface ChatExecutionService {
+  deleteChatQueueFile(chatId: string): Promise<void>;
+  submit(chatId: string, command: string, options: RunAgentTurnOptions): Promise<void>;
+  registerPendingUserInput(
+    chatId: string,
+    command: string,
+    options: PendingUserInputRegistrationOptions,
+  ): Promise<void>;
+  reserveDirectTurn(chatId: string, turn?: TurnIdentity): DirectTurnReservation;
+  assertDirectTurnReservationActive(reservation: DirectTurnReservation): void;
+  consumeRecoveredInputContinuationForDirectTurn(
+    reservation: DirectTurnReservation,
+  ): Promise<StoredChatExecutionControlState>;
+  releaseDirectTurn(reservation: DirectTurnReservation): Promise<void>;
+  completeDirectTurn(reservation: DirectTurnReservation): Promise<void>;
+  failDirectTurn(reservation: DirectTurnReservation): Promise<void>;
+  runReservedTurn(
+    reservation: DirectTurnReservation,
+    command: string,
+    options: RunAgentTurnOptions,
+  ): Promise<void>;
+  stopActiveTurn(chatId: string): Promise<StopActiveTurnResult>;
+  interruptActiveTurn(chatId: string): Promise<boolean>;
+  abortForChatDeletion(chatId: string): Promise<boolean>;
+  beginShutdown(): string[];
+  abortForShutdown(chatId: string): Promise<boolean>;
+  waitForExecutionOwners(): Promise<void>;
+  getQueuedTurnFinalization(
+    chatId: string,
+    turnId: string | undefined,
+  ): Promise<QueuedTurnFinalizationOutcome> | null;
+  isChatDraining(chatId: string): boolean;
+  isChatExecutionReserved(chatId: string): boolean;
+  hasChatExecutionOwner(chatId: string): boolean;
+  triggerDrain(chatId: string): Promise<void>;
+  readChatExecutionControl(chatId: string): Promise<StoredChatExecutionControlState>;
+  hasAppliedQueueCreateCommand(chatId: string, commandKey: string, entryId: string): Promise<boolean>;
+  createChatQueueEntry(
+    chatId: string,
+    content: string,
+    command?: QueueCommandIdentity,
+    receipts?: ReceiptRetention,
+  ): Promise<QueueCommandMutationResult & { entry: QueueEntry | null }>;
+  replaceChatQueueEntry(
+    chatId: string,
+    entryId: string,
+    content: string,
+    expectedRevision: number,
+    command?: QueueCommandIdentity,
+    receipts?: ReceiptRetention,
+  ): Promise<QueueCommandMutationResult & { entry: QueueEntry | null }>;
+  deleteChatQueueEntry(
+    chatId: string,
+    entryId: string,
+    command?: QueueCommandIdentity,
+    receipts?: ReceiptRetention,
+  ): Promise<QueueCommandMutationResult>;
+  deliverActiveInput(
+    chatId: string,
+    content: string,
+    options?: RunAgentTurnOptions,
+    afterPendingRegistered?: () => Promise<void>,
+  ): Promise<boolean>;
+  clearChatQueue(chatId: string): Promise<StoredChatExecutionControlState>;
+  pauseChatQueue(chatId: string): Promise<StoredChatExecutionControlState>;
+  resumeChatQueue(chatId: string, pauseId: string): Promise<StoredChatExecutionControlState>;
+  continuePastRecoveredInput(
+    chatId: string,
+    continuationId: string,
+  ): Promise<StoredChatExecutionControlState>;
+  dropRecoveredInputContinuation(chatId: string): Promise<StoredChatExecutionControlState>;
+  requeueAndPauseChat(
+    chatId: string,
+    entryId: string,
+    kind: AutomaticQueuePauseKind,
+  ): Promise<StoredChatExecutionControlState>;
+}
+
+export const EMPTY_RECEIPT_RETENTION: ReceiptRetention = { protectedKeys: new Set() };
+
+export function transitionContext(): TransitionContext {
+  return { now: new Date().toISOString(), newId: () => crypto.randomUUID() };
+}
+
+export function executionTurnIdentity(turn: TurnIdentity): TurnIdentity | undefined {
+  if (!turn.turnId && !turn.clientRequestId) return undefined;
+  return {
+    ...(turn.turnId ? { turnId: turn.turnId } : {}),
+    ...(turn.clientRequestId ? { clientRequestId: turn.clientRequestId } : {}),
+  };
+}
+
+export function transitionError(
+  rejection: TransitionRejection,
+  control: StoredChatExecutionControlState,
+): DomainError {
+  switch (rejection.code) {
+    case 'QUEUE_ENTRY_NOT_FOUND':
+      return new QueueEntryMutationError(
+        rejection.code,
+        'This queued message is no longer available',
+        control,
+      );
+    case 'QUEUE_ENTRY_ALREADY_SENT':
+      return new QueueEntryMutationError(
+        rejection.code,
+        'This queued message has already been sent',
+        control,
+      );
+    case 'QUEUE_ENTRY_REVISION_CONFLICT':
+      return new QueueEntryMutationError(
+        rejection.code,
+        'This queued message changed before it could be saved',
+        control,
+      );
+    case 'QUEUE_PAUSE_CHANGED':
+      return new QueuePauseChangedError(control);
+    case 'RECOVERED_INPUT_CONTINUATION_CHANGED':
+      return new RecoveredInputContinuationChangedError(control);
+    case 'RECOVERED_INPUT_CONTINUATION_REQUIRES_QUEUE':
+      return new RecoveredInputContinuationRequiresQueueError(control);
+  }
+}
